+++ math/src/tuple/vector/mod.rs
@@ -84,7 +84,7 @@ impl std::fmt::Display for Vector {
 #[macro_export]
 macro_rules! vector {
     ($x:expr, $y:expr, $z:expr) => {
-        Vector::vector($x, $y, $z)
+        Vector::vector($x as f32, $y as f32, $z as f32)
     };
 }
